HAL: linux: Set socket protocol only for receive sockets
When creating raw sockets the protocol must be set to 0.

In case for receive sockets this means we will receive no packets until
we "bind" the socket with a non-zero protocol.

In case for only transmit sockets this avoids an expensive call packet
receiving function in kernel thats overflow internal receive socket
buffer.